requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
DateTime==4.3
defusedxml==0.7.1
Django==3.2
django-allauth==0.42.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.0
joblib==1.0.1
Markdown==3.3.4
mysqlclient==2.0.3
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.4
PyJWT==2.0.1
pyrandom==0.0.1
python-dateutil==2.8.1
python-dotenv==0.17.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
uuid==1.30
zipp==3.4.1
zope.interface==5.3.0
